requirements.txt
django==1.9.1
djangorestframework==3.5.3
pygments==2.2.0
python-dateutil==2.6.0